// features/api/model/usersIndex200.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */
import type { UserResource } from './userResource';
import type { UsersIndex200Links } from './usersIndex200Links';
import type { UsersIndex200Meta } from './usersIndex200Meta';

export type UsersIndex200 = {
  data: UserResource[];
  links: UsersIndex200Links;
  meta: UsersIndex200Meta;
};